backend/requirements.txt
aniso8601==3.0.0
astroid==1.6.4
dj-database-url==0.5.0
Django==2.0.5
django-authtools==1.6.0
django-cors-headers==2.2.0
django-environ==0.4.4
django-graphql-jwt==0.1.10
graphene==2.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
promise==2.1
psycopg2==2.7.4
PyJWT==1.6.1
pylint==1.9.1
python-decouple==3.1
pytz==2018.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
wrapt==1.10.11
